+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/revisitor/logo_lang/revisitor-comp/test/impl/operation/impl/ForwardOpImpl.java
@@ -1,5 +1,6 @@
 package test.impl.operation.impl;
 
+import java.lang.Double;
 import kmLogo.Expression;
 import kmLogo.Forward;
 import kmLogo.Turtle;
@@ -60,7 +61,7 @@ public class ForwardOpImpl extends PrimitiveOpImpl implements ForwardOp {
     double result;
     double move = ((double) (rev.$((Expression)this.obj.getSteps()).eval(((Turtle) (turtle)))));
     rev.$((Turtle)turtle).forward(((double) (move)));
-    result = 0.0;
+    result = ((Double) 0.0);
     return result;
   }
 }
